+++ src/brickd/stack.h
@@ -31,7 +31,7 @@ typedef struct _Stack Stack;
 
 typedef struct {
 	uint32_t uid; // always little endian
-	int opaque;
+	uint64_t opaque;
 } Recipient;
 
 typedef int (*StackDispatchRequestFunction)(Stack *stack, Packet *request, Recipient *recipient);
@@ -48,7 +48,7 @@ int stack_create(Stack *stack, const char *name,
                  StackDispatchRequestFunction dispatch_request);
 void stack_destroy(Stack *stack);
 
-int stack_add_recipient(Stack *stack, uint32_t uid /* always little endian */, int opaque);
+int stack_add_recipient(Stack *stack, uint32_t uid /* always little endian */, uint64_t opaque);
 Recipient *stack_get_recipient(Stack *stack, uint32_t uid /* always little endian */);
 
 int stack_dispatch_request(Stack *stack, Packet *request, bool force);
